feat(web): comprehensive seo improvements
- add canonical urls, hreflang (x-default), og/twitter metadata to all pages
- add json-ld structured data: organization, website, software, faqpage, blogging
- convert client pages to server wrappers for metadata export support
- fix sitemap: remove auth/glossary pages, add pricing/security, dynamic blog posts
- extend robots.txt disallow list for auth and api routes
- add aria-label to h1 for typewriter animation seo fix
- delete unused glossary and design-system pages
- add server-side redirect on root page
- remove unused @vm0/ui dependency
