+++ turbo/apps/web/app/page.tsx
@@ -1,11 +1,5 @@
-// Root page — redirects to the default locale
+import { redirect } from "next/navigation";
+
 export default function RootRedirect() {
-  return (
-    <div style={{ padding: "50px", textAlign: "center" }}>
-      <p>Redirecting to /en...</p>
-      <script
-        dangerouslySetInnerHTML={{ __html: `window.location.href = '/en';` }}
-      />
-    </div>
-  );
+  redirect("/en");
 }
