+++ turbo/apps/web/app/sign-in/[[...sign-in]]/SignInClient.tsx
@@ -1,0 +1,16 @@
+"use client";
+
+import { SignIn } from "@clerk/nextjs";
+import { useTheme } from "../../components/ThemeProvider";
+import { AuthLayout } from "../../components/auth/AuthLayout";
+import { getClerkAppearance } from "../../components/auth/clerk-appearance";
+
+export default function SignInClient() {
+  const { theme } = useTheme();
+
+  return (
+    <AuthLayout>
+      <SignIn appearance={getClerkAppearance(theme)} />
+    </AuthLayout>
+  );
+}
