+++ turbo/apps/web/app/sign-up/[[...sign-up]]/SignUpClient.tsx
@@ -1,0 +1,16 @@
+"use client";
+
+import { SignUp } from "@clerk/nextjs";
+import { useTheme } from "../../components/ThemeProvider";
+import { AuthLayout } from "../../components/auth/AuthLayout";
+import { getClerkAppearance } from "../../components/auth/clerk-appearance";
+
+export default function SignUpClient() {
+  const { theme } = useTheme();
+
+  return (
+    <AuthLayout>
+      <SignUp appearance={getClerkAppearance(theme)} />
+    </AuthLayout>
+  );
+}
