+++ turbo/apps/web/package.json
@@ -34,7 +34,6 @@
     "@ts-rest/core": "3.53.0-rc.1",
     "@ts-rest/serverless": "3.53.0-rc.1",
     "@vm0/core": "workspace:*",
-    "@vm0/ui": "workspace:*",
     "ably": "^2.17.0",
     "archiver": "^7.0.1",
     "cpu-features": "0.0.10",
